Crate utc2tt

This binary crate converts datetimes from UTC to TT.

For example:

$ utc2tt 2016-12-31T23:59:60
2017-01-01T00:01:08.184

As you can see from the above example, it takes leap seconds into account.

In this execution, it assume that TT = TAI + 32.184.

Arguments

It takes one or more datetimes as argument.

$ utc2tt 2016-12-31T23:59:59 2016-12-31T23:59:60 2017-01-01T00:00:00
2017-01-01T00:01:07.184
2017-01-01T00:01:08.184
2017-01-01T00:01:09.184

Instead of specifying the datetimes as arguments, they can be specified from the standard input.

Options

--dt-fmt <dt_fmt>

format of input/output datetimes. For example:

$ utc2tt --dt-fmt %Y%m%d%H%M%S%.3f 20161231235960.000
20170101000108.184

If both of this option and an environment variable DT_FMT are not specified, it uses the default: %Y-%m-%dT%H:%M:%S%.3f.

-H, --io-pair

If it is specified, not only converted datetime but also input datetime are output. For example:

$ utc2tt -H 2016-12-31T23:59:59 2016-12-31T23:59:60 2017-01-01T00:00:00
2016-12-31T23:59:59 2017-01-01T00:01:07.184
2016-12-31T23:59:60 2017-01-01T00:01:08.184
2017-01-01T00:00:00 2017-01-01T00:01:09.184

--tai-utc-table <tai_utc_table_file>

It specifies a file which contains definition of TAI-UTC differences. If you use it, the option --tai-utc-table-dt-fmt may be useful.

If both of this option and an environment variable TAI_UTC_TABLE are not specified, it uses the default: tai-utc.txt in directory of executable file. If the default file also does not exist, use the built-in table in the program.

--tai-utc-table-dt-fmt <tai_utc_table_dt_fmt>

format of datetimes in TAI-UTC table file. It will be used when you use your TAI-UTC table file with the option --tai-utc-table.

If both of this option and an environment variable TAI_UTC_TABLE_DT_FMT are not specified, it uses the default: %Y-%m-%dT%H:%M:%S%.3f.

-V, --version

Show the version of this executable.

-h, --help

Show help document of this executable.

Environment variables

It looks for below environment variables.

DT_FMT

Look for a description for an option --dt-fmt.

TAI_UTC_TABLE

Look for a description for an option --tai-utc-table.

TAI_UTC_TABLE_DT_FMT

Look for a description for an option --tai-utc-table-dt-fmt.

Standard input

If the datetimes are not specified as arguments, they can be entered from the standard input instead.